REFACTOR: move `BinOpSet` interface
